renamed my local asc() and chr() functions
This library isn't very important now that
OpenSCAD can display text natively.  However,
I recently tried to use it on a small project
and it broke everything.  What happened is that
since I wrote this module, OpenSCAD introduced
a built-in function called `chr`, which I was
overwriting in `ascii.scad`.  I renamed the
function name, and `asc` too, just to be careful.

* `asc()` ==> `myasc()`
* `chr()` ==> `mychr()`